+++ conf/conf_test.go
@@ -1,0 +1,13 @@
+package conf
+
+import "testing"
+
+func TestConf(t *testing.T) {
+	// TODO: Lookup
+	// TODO: Create
+	// TODO: getKey
+	// TODO: uintOrDefault
+	// TODO: intOrDefault
+	// TODO: stringOrDefault
+	// TODO: stringListOrDefault
+}