+++ sys/sys_test.go
@@ -1,0 +1,9 @@
+package sys
+
+import "testing"
+
+func TestEnv(t *testing.T) {
+	// TODO: IsValidEnv
+	// TODO: IsProdEnv
+	// TODO: RunInitializers
+}